+++ languagetool-commandline/src/main/java/org/languagetool/commandline/CommandLineOptions.java
@@ -67,10 +67,6 @@ public class CommandLineOptions {
   private Language motherTongue = null;
   @Nullable
   private File languageModel = null;
-  @Nullable
-  private File word2vecModel = null;
-  @Nullable
-  private File neuralNetworkModel = null;
 
   @Nullable
   private File fasttextModel = null;
@@ -257,37 +253,6 @@ public class CommandLineOptions {
    */
   public void setLanguageModel(File languageModel) {
     this.languageModel = languageModel;
-  }
-
-  /**
-   * @return a directory with a word2vec language model for use with neural network rules in sub directories like ({@code en}), or {@code null}
-   * @since 4.0
-   */
-  @Nullable
-  public File getWord2VecModel() {
-    return word2vecModel;
-  }
-
-  /**
-   * @since 4.0
-   */
-  public void setWord2VecModel(File neuralNetworkLanguageModel) {
-    this.word2vecModel = neuralNetworkLanguageModel;
-  }
-
-  /**
-   * @since 4.4
-   */
-  @Nullable
-  public File getNeuralNetworkModel() {
-    return neuralNetworkModel;
-  }
-
-  /**
-   * @since 4.4
-   */
-  public void setNeuralNetworkModel(File neuralNetworkModel) {
-    this.neuralNetworkModel = neuralNetworkModel;
   }
 
   /**