+++ languagetool-commandline/src/main/java/org/languagetool/commandline/Main.java
@@ -74,12 +74,6 @@ class Main {
     if (options.getLanguageModel() != null) {
       lt.activateLanguageModelRules(options.getLanguageModel());
     }
-    if (options.getWord2VecModel() != null) {
-      lt.activateWord2VecModelRules(options.getWord2VecModel());
-    }
-    if (options.getNeuralNetworkModel() != null) {
-      lt.activateNeuralNetworkRules(options.getNeuralNetworkModel());
-    }
     lt.activateRemoteRules(options.getRemoteRulesFile() != null ? new File(options.getRemoteRulesFile()) : null);
     Tools.selectRules(lt, options.getDisabledCategories(), options.getEnabledCategories(),
             new HashSet<>(options.getDisabledRules()), new HashSet<>(options.getEnabledRules()), options.isUseEnabledOnly(), options.isEnableTempOff());