+++ languagetool-core/src/main/java/org/languagetool/Language.java
@@ -26,7 +26,6 @@ import org.languagetool.language.Contributor;
 import org.languagetool.languagemodel.LanguageModel;
 import org.languagetool.languagemodel.LuceneLanguageModel;
 import org.languagetool.rules.*;
-import org.languagetool.rules.neuralnetwork.Word2VecModel;
 import org.languagetool.rules.patterns.AbstractPatternRule;
 import org.languagetool.rules.patterns.PatternRuleLoader;
 import org.languagetool.rules.patterns.Unifier;
@@ -244,34 +243,6 @@ public abstract class Language {
     ResourceBundle messageBundle, List<RemoteRuleConfig> configs, UserConfig userConfig,
     Language motherTongue, List<Language> altLanguages, boolean inputLogging) throws IOException {
     return Function.identity();
-  }
-
-  /**
-   * @param indexDir directory with a subdirectories like 'en', each containing dictionary.txt and final_embeddings.txt
-   * @return a {@link Word2VecModel} or {@code null} if this language doesn't support one
-   * @since 4.0
-   */
-  @Nullable
-  public Word2VecModel getWord2VecModel(File indexDir) throws IOException {
-    return null;
-  }
-
-  /**
-   * Get a list of rules that require a {@link Word2VecModel}. Returns an empty list for
-   * languages that don't have such rules.
-   * @since 4.0
-   */
-  public List<Rule> getRelevantWord2VecModelRules(ResourceBundle messages, Word2VecModel word2vecModel) throws IOException {
-    return Collections.emptyList();
-  }
-
-  /**
-   * Get a list of rules that load trained neural networks. Returns an empty list for
-   * languages that don't have such rules.
-   * @since 4.4
-   */
-  public List<Rule> getRelevantNeuralNetworkModels(ResourceBundle messages, File modelDir) {
-    return Collections.emptyList();
   }
 
   /**