+++ languagetool-gui-commons/src/main/java/org/languagetool/gui/Configuration.java
@@ -92,7 +92,6 @@ public class Configuration {
   private static final String MOTHER_TONGUE_KEY = "motherTongue";
   private static final String FIXED_LANGUAGE_KEY = "fixedLanguage";
   private static final String NGRAM_DIR_KEY = "ngramDir";
-  private static final String WORD2VEC_DIR_KEY = "word2vecDir";
   private static final String AUTO_DETECT_KEY = "autoDetect";
   private static final String TAGGER_SHOWS_DISAMBIG_LOG_KEY = "taggerShowsDisambigLog";
   private static final String SERVER_RUN_KEY = "serverMode";
@@ -174,7 +173,6 @@ public class Configuration {
   private Language motherTongue = null;
   private Language fixedLanguage = null;
   private File ngramDirectory;
-  private File word2vecDirectory;
   private boolean runServer;
   private boolean autoDetect;
   private boolean taggerShowsDisambigLog;
@@ -249,7 +247,6 @@ public class Configuration {
   /**
    * Initialize variables and clears Maps, Sets and Lists
    */
-  
   public void initOptions() {
     configForOtherLanguages.clear();
     underlineColors.clear();
@@ -266,7 +263,6 @@ public class Configuration {
 
     enabledRulesOnly = false;
     ngramDirectory = null;
-    word2vecDirectory = null;
     runServer = false;
     autoDetect = false;
     taggerShowsDisambigLog = false;
@@ -320,7 +316,6 @@ public class Configuration {
     this.motherTongue = configuration.motherTongue;
     this.fixedLanguage = configuration.fixedLanguage;
     this.ngramDirectory = configuration.ngramDirectory;
-    this.word2vecDirectory = configuration.word2vecDirectory;
     this.runServer = configuration.runServer;
     this.autoDetect = configuration.autoDetect;
     this.taggerShowsDisambigLog = configuration.taggerShowsDisambigLog;
@@ -879,23 +874,6 @@ public class Configuration {
   }
 
   /**
-   * Directory with word2vec data or null.
-   * @since 4.0
-   */
-  @Nullable
-  public File getWord2VecDirectory() {
-    return word2vecDirectory;
-  }
-
-  /**
-   * Sets the directory with word2vec data (may be null).
-   * @since 4.0
-   */
-  public void setWord2VecDirectory(File dir) {
-    this.word2vecDirectory = dir;
-  }
-
-  /**
    * @since 2.8
    */
   public Map<ITSIssueType, Color> getErrorColors() {
@@ -1249,10 +1227,6 @@ public class Configuration {
     if (ngramDir != null) {
       ngramDirectory = new File(ngramDir);
     }
-    String word2vecDir = (String) props.get(prefix + WORD2VEC_DIR_KEY);
-    if (word2vecDir != null) {
-      word2vecDirectory = new File(word2vecDir);
-    }
 
     autoDetect = "true".equals(props.get(prefix + AUTO_DETECT_KEY));
     taggerShowsDisambigLog = "true".equals(props.get(prefix + TAGGER_SHOWS_DISAMBIG_LOG_KEY));
@@ -1554,7 +1528,6 @@ public class Configuration {
     allProfileKeys.add(LANGUAGE_KEY);
     allProfileKeys.add(FIXED_LANGUAGE_KEY);
     allProfileKeys.add(NGRAM_DIR_KEY);
-    allProfileKeys.add(WORD2VEC_DIR_KEY);
     allProfileKeys.add(AUTO_DETECT_KEY);
     allProfileKeys.add(TAGGER_SHOWS_DISAMBIG_LOG_KEY);
     allProfileKeys.add(SERVER_RUN_KEY);
@@ -1649,9 +1622,6 @@ public class Configuration {
     }
     if (ngramDirectory != null) {
       props.setProperty(prefix + NGRAM_DIR_KEY, ngramDirectory.getAbsolutePath());
-    }
-    if (word2vecDirectory != null) {
-      props.setProperty(prefix + WORD2VEC_DIR_KEY, word2vecDirectory.getAbsolutePath());
     }
     props.setProperty(prefix + AUTO_DETECT_KEY, Boolean.toString(autoDetect));
     props.setProperty(prefix + TAGGER_SHOWS_DISAMBIG_LOG_KEY, Boolean.toString(taggerShowsDisambigLog));